% Auto-generated: do not edit by hand
\name{'dc'Datepicker}

\alias{'dc'Datepicker}

\title{Datepicker component}

\description{

}

\usage{
'dc'Datepicker(id=NULL, label=NULL, initialDate=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{initialDate}{Character. The initialDate controls the calendar's inistal date. like 2012/06/01}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

